rename ratelimiting_sampler.js -> rate_limiting_sampler.js
